requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
beautifulsoup4==4.12.2
# Optional for better recognition (large)
transformers==4.44.0
torch  # install appropriate torch wheel for your system (see README_RUN for guidance)
